backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
pypdf==5.9.0
reportlab==4.4.3
python-docx==1.2.0
pytesseract==0.3.13
Pillow==11.3.0
requests==2.32.5
